@@ -1,170 +1,160 @@
 {
-  "name": "equicord-companion",
-  "displayName": "Equicord Companion",
-  "version": "0.2.0",
-  "description": "Equicord Companion is a vscode extension to test Equicord patches & webpack finds right from the comfort of your IDE",
-  "categories": [
-    "Other",
-    "Snippets",
-    "Testing"
-  ],
-  "keywords": [
-    "Discord",
-    "Equicord",
-    "Modding",
-    "Typescript"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Equicord/Companion"
-  },
-  "license": "GPL-3.0",
-  "author": "Thororen <78185467+thororen1234@users.noreply.github.com>",
-  "contributors": [
-    {
-      "name": "Vendicated",
-      "url": "https://vendicated.dev/"
-    },
-    {
-      "name": "sadan",
-      "url": "https://sadan.zip"
-    }
-  ],
-  "publisher": "thororen",
-  "main": "./dist/extension.js",
-  "scripts": {
-    "build": "node ./scripts/build.mjs",
-    "watch": "node ./scripts/watch.mjs",
-    "lint": "eslint && pnpm typecheck",
-    "package": "vsce package --no-dependencies",
-    "vscode:prepublish": "pnpm build && pnpm lint",
-    "publish": "vsce publish --no-dependencies",
-    "typecheck": "tsc --noEmit",
-    "postinstall": "(cd src/webview && pnpm i)"
-  },
-  "contributes": {
-    "configuration": {
-      "title": "Equicord Companion",
-      "properties": {
-        "equicord-companion.showSidebar": {
-          "default": false,
-          "description": "Show the sidebar. This is intened to be used by workspaces in their .vscode/settings.json file, use alwaysShowSidebar instead to override this setting",
-          "type": "boolean"
-        },
-        "equicord-companion.alwaysShowSidebar": {
-          "default": false,
-          "description": "Always show the sidebar",
-          "type": "boolean"
-        }
-      }
-    },
-    "viewsContainers": {
-      "activitybar": [
-        {
-          "id": "equicord-companion-panel",
-          "title": "Companion",
-          "icon": "assets/oneko.svg"
-        }
-      ]
-    },
-    "views": {
-      "equicord-companion-panel": [
-        {
-          "id": "equicordSettings",
-          "name": "Settings",
-          "when": "config.equicord-companion.showSidebar || config.equicord-companion.alwaysShowSidebar",
-          "visibility": "visible"
-        }
-      ]
-    },
-    "commands": [
-      {
-        "command": "equicord-companion.runReporter",
-        "title": "Run Reporter"
-      },
-      {
-        "command": "equicord-companion.diffModuleSearch",
-        "title": "Diff Module Search"
-      },
-      {
-        "command": "equicord-companion.diffModule",
-        "title": "Diff Module"
-      },
-      {
-        "command": "equicord-companion.extract",
-        "title": "Extract"
-      },
-      {
-        "command": "equicord-companion.extractSearch",
-        "title": "Extract With Search"
-      },
-      {
-        "command": "equicord-companion.cacheModules",
-        "title": "Cache Modules"
-      },
-      {
-        "command": "equicord-companion.testDeps",
-        "title": "Test Dependencies"
-      },
-      {
-        "command": "equicord-companion.initDeps",
-        "title": "init deps"
-      }
+    "name": "equicord-companion",
+    "displayName": "Equicord Companion",
+    "version": "2.5.2025",
+    "description": "Equicord Companion is a vscode extension to test Equicord patches & webpack finds right from the comfort of your IDE",
+    "categories": [
+        "Other",
+        "Snippets",
+        "Testing"
     ],
-    "snippets": [
-      {
-        "language": "typescript",
-        "path": "./snippets/ts.json"
-      },
-      {
-        "language": "typescriptreact",
-        "path": "./snippets/ts.json"
-      },
-      {
-        "language": "typescriptreact",
-        "path": "./snippets/tsx.json"
-      }
-    ]
-  },
-  "activationEvents": [
-    "onLanguage:typescript",
-    "onLanguage:typescriptreact"
-  ],
-  "dependencies": {
-    "prettier": "^3.3.3",
-    "tsutils": "^3.21.0",
-    "typescript": "^4.9.5",
-    "ws": "^8.12.1"
-  },
-  "devDependencies": {
-    "@intrnl/xxhash64": "^0.1.2",
-    "@stylistic/eslint-plugin": "^2.7.2",
-    "@types/node": "16.x",
-    "@types/vscode": "^1.70.0",
-    "@types/ws": "^8.5.4",
-    "@typescript-eslint/eslint-plugin": "^5.53.0",
-    "@typescript-eslint/parser": "^5.53.0",
-    "@vscode/vsce": "^2.18.0",
-    "esbuild": "^0.23.1",
-    "eslint": "^9.9.1",
-    "eslint-plugin-simple-import-sort": "^12.1.1",
-    "nanoid": "^5.0.9",
-    "typescript-eslint": "^8.3.0"
-  },
-  "engines": {
-    "vscode": "^1.70.0"
-  },
-  "icon": "assets/icon.png",
-  "galleryBanner": {
-    "color": "#4b4b4b",
-    "theme": "dark"
-  },
-  "extensionKind": [
-    "ui"
-  ],
-  "sponsor": {
-    "url": "https://github.com/sponsors/verticalsync"
-  },
-  "homepage": "https://github.com/Equicord/Companion",
-  "packageManager": "pnpm@9.1.4+sha512.9df9cf27c91715646c7d675d1c9c8e41f6fce88246f1318c1aa6a1ed1aeb3c4f032fcdf4ba63cc69c4fe6d634279176b5358727d8f2cc1e65b65f43ce2f8bfb0"
+    "keywords": [
+        "Discord",
+        "Equicord",
+        "Modding",
+        "Typescript"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/Equicord/Companion"
+    },
+    "license": "GPL-3.0",
+    "author": "Thororen <78185467+thororen1234@users.noreply.github.com>",
+    "contributors": [
+        {
+            "name": "Vendicated",
+            "url": "https://vendicated.dev/"
+        },
+        {
+            "name": "sadan",
+            "url": "https://sadan.zip"
+        }
+    ],
+    "publisher": "thororen",
+    "main": "./dist/extension.js",
+    "scripts": {
+        "build": "node ./scripts/build.mjs",
+        "watch": "node ./scripts/watch.mjs",
+        "lint": "eslint && pnpm typecheck",
+        "package": "vsce package --no-dependencies",
+        "vscode:prepublish": "pnpm build && pnpm lint",
+        "publish": "vsce publish --no-dependencies",
+        "typecheck": "tsc --noEmit",
+        "postinstall": "(cd src/webview && pnpm i)"
+    },
+    "contributes": {
+        "configuration": {
+            "title": "Equicord Companion",
+            "properties": {
+                "equicord-companion.showSidebar": {
+                    "default": false,
+                    "description": "Show the sidebar. This is intened to be used by workspaces in their .vscode/settings.json file, use alwaysShowSidebar instead to override this setting",
+                    "type": "boolean"
+                },
+                "equicord-companion.alwaysShowSidebar": {
+                    "default": false,
+                    "description": "Always show the sidebar",
+                    "type": "boolean"
+                }
+            }
+        },
+        "viewsContainers": {
+            "activitybar": [
+                {
+                    "id": "equicord-companion-panel",
+                    "title": "Companion",
+                    "icon": "assets/icon.svg"
+                }
+            ]
+        },
+        "views": {
+            "equicord-companion-panel": [
+                {
+                    "id": "equicordSettings",
+                    "name": "Settings",
+                    "when": "config.equicord-companion.showSidebar || config.equicord-companion.alwaysShowSidebar",
+                    "visibility": "visible"
+                }
+            ]
+        },
+        "commands": [
+            {
+                "command": "equicord-companion.runReporter",
+                "title": "Run Reporter"
+            },
+            {
+                "command": "equicord-companion.diffModuleSearch",
+                "title": "Diff Module Search"
+            },
+            {
+                "command": "equicord-companion.diffModule",
+                "title": "Diff Module"
+            },
+            {
+                "command": "equicord-companion.extract",
+                "title": "Extract"
+            },
+            {
+                "command": "equicord-companion.extractSearch",
+                "title": "Extract With Search"
+            }
+        ],
+        "snippets": [
+            {
+                "language": "typescript",
+                "path": "./snippets/ts.json"
+            },
+            {
+                "language": "typescriptreact",
+                "path": "./snippets/ts.json"
+            },
+            {
+                "language": "typescriptreact",
+                "path": "./snippets/tsx.json"
+            }
+        ]
+    },
+    "activationEvents": [
+        "onLanguage:typescript",
+        "onLanguage:typescriptreact"
+    ],
+    "dependencies": {
+        "@intrnl/xxhash64": "^0.1.2",
+        "fast-diff": "^1.3.0",
+        "nanoid": "^5.0.9",
+        "prettier": "^3.4.2",
+        "tsutils": "^3.21.0",
+        "typescript": "^5.7.3",
+        "ws": "^8.18.0"
+    },
+    "devDependencies": {
+        "@stylistic/eslint-plugin": "^3.0.1",
+        "@types/node": "~22.13.1",
+        "@types/vscode": "^1.96.0",
+        "@types/ws": "^8.5.14",
+        "@typescript-eslint/eslint-plugin": "^8.23.0",
+        "@typescript-eslint/parser": "^8.23.0",
+        "@vscode/vsce": "^3.2.2",
+        "esbuild": "^0.24.2",
+        "eslint": "^9.19.0",
+        "eslint-plugin-simple-import-sort": "^12.1.1",
+        "eslint-plugin-unused-imports": "^4.1.4",
+        "typescript-eslint": "^8.23.0"
+    },
+    "engines": {
+        "vscode": "^1.70.0"
+    },
+    "icon": "assets/icon.png",
+    "galleryBanner": {
+        "color": "#4b4b4b",
+        "theme": "dark"
+    },
+    "extensionKind": [
+        "ui"
+    ],
+    "sponsor": {
+        "url": "https://github.com/sponsors/verticalsync"
+    },
+    "homepage": "https://github.com/Equicord/Companion",
+    "packageManager": "pnpm@9.1.4+sha512.9df9cf27c91715646c7d675d1c9c8e41f6fce88246f1318c1aa6a1ed1aeb3c4f032fcdf4ba63cc69c4fe6d634279176b5358727d8f2cc1e65b65f43ce2f8bfb0"
 }
