@@ -1,16 +1,25 @@
-import { formatModule, mkStringUri, sendAndGetData } from "@server/webSocketServer";
-import { assert, dir } from "console";
-import { ModuleCache, ModuleDepManager } from "modules/cache";
 import {
-    collectVariableUsage,
-    getTokenAtPosition,
-    VariableInfo,
-} from "tsutils";
+    findObjectLiteralByKey,
+    findParrent,
+    findReturnIdentifier,
+    findReturnPropertyAccessExpression,
+    findWebpackArg,
+    getLeadingIdentifier,
+    getModuleId,
+    makeRange,
+    zeroRange
+} from "@ast/util";
+import { ModuleCache, ModuleDepManager } from "@modules/cache";
+import { format } from "@modules/format";
+import { formatModule, mkStringUri, sendAndGetData } from "@server/index";
+import { Definitions, ExportMap, ModuleDeps, References } from "@type/ast";
+
+import { collectVariableUsage, getTokenAtPosition, VariableInfo } from "tsutils";
 import {
-    BinaryExpression,
     CallExpression,
     createSourceFile,
     Identifier,
+    isArrowFunction,
     isBinaryExpression,
     isCallExpression,
     isFunctionDeclaration,
@@ -27,81 +36,11 @@ import {
     ScriptKind,
     ScriptTarget,
     SourceFile,
-    StringLiteral,
-    VariableDeclaration,
 } from "typescript";
-import * as vscode from "vscode";
-
-import format from "../format";
-import {
-    findObjectLiteralByKey,
-    findParrent,
-    findReturnIdentifier,
-    findReturnPropertyAccessExpression,
-    findWebpackArg,
-    getLeadingIdentifier,
-    getModuleId,
-    makeRange,
-    zeroRange,
-} from "./util";
-
-type Definitions = Promise<
-    vscode.Definition | vscode.LocationLink[] | null | undefined
->;
-type References = Promise<vscode.Location[] | null | undefined>;
-export class ReferenceProvider implements vscode.ReferenceProvider {
-    async provideReferences(document: vscode.TextDocument, position: vscode.Position, context: vscode.ReferenceContext, token: vscode.CancellationToken): References {
-        const y = require("path");
-        if (!await ModuleCache.hasCache()) {
-            vscode.window.showErrorMessage("No Module Cache found, please download modules first");
-            return;
-        }
-        if (!ModuleDepManager.hasModDeps()) {
-            await ModuleDepManager.initModDeps({
-                fromDisk: true
-            });
-        }
-        try {
-            return await new WebpackAstParser(document.getText()).generateReferences(document, position);
-        } catch (e) {
-            console.error(e);
-        }
-    }
-
-}
-export class DefinitionProvider implements vscode.DefinitionProvider {
-    async provideDefinition(
-        document: vscode.TextDocument,
-        position: vscode.Position
-    ): Definitions {
-        try {
-            const text = document.getText();
-            // not sure if substring is a good idea here
-            // just dont want to search really long webpack modules
-            if (
-                !text.startsWith("//WebpackModule") ||
-                text.substring(0, 100).includes("//OPEN FULL MODULE:")
-            )
-                return;
-            return await new WebpackAstParser(text).generateDefinitions(
-                document,
-                position
-            );
-        } catch (e) {
-            console.error(e);
-        }
-    }
-}
-interface ModuleDeps {
-    lazy: string[];
-    sync: string[];
-}
-interface ExportMap {
-    // ranges of code that will count as references to this export
-    [exposedName: string]: (vscode.Range | undefined)[];
-}
+import { Location, Position, Range, TextDocument } from "vscode";
 
 // FIXME: rewrite to use module cache
+
 export class WebpackAstParser {
     private text: string;
     private sourceFile: SourceFile;
@@ -113,8 +52,14 @@ export class WebpackAstParser {
     private uses: VariableInfo | undefined;
     private thisId: string | null;
 
-    public constructor(text: string) {
-        this.text = text;
+    public constructor(text: string | { document: TextDocument; } | TextDocument) {
+        if (typeof text === "string") {
+            this.text = text;
+        } else if ("document" in text) {
+            this.text = text.document.getText();
+        } else {
+            this.text = text.getText();
+        }
         this.sourceFile = createSourceFile(
             "module.js",
             this.text,
@@ -155,7 +100,7 @@ export class WebpackAstParser {
          * things like wreq(moduleid)
          */
         const wreqCalls = this.uses.uses.map(x => x.location).flatMap(v => {
-            const p = findParrent<CallExpression>(v, n => isCallExpression(n));
+            const p = findParrent(v, isCallExpression);
             if (!p || p.expression !== v) return [];
 
             if (p.arguments.length === 1 && isNumericLiteral(p.arguments[0]))
@@ -166,7 +111,7 @@ export class WebpackAstParser {
         const lazyModules = this.uses.uses.map(x => x.location).flatMap(v => {
             const [, prop] = getLeadingIdentifier(v);
             if (prop?.text !== "bind") return [];
-            const call = findParrent<CallExpression>(v, n => isCallExpression(n));
+            const call = findParrent(v, isCallExpression);
             if (!call) return [];
 
             if (call.arguments.length === 2 && isNumericLiteral(call.arguments[1]))
@@ -181,8 +126,8 @@ export class WebpackAstParser {
     }
 
     public async generateDefinitions(
-        document: vscode.TextDocument,
-        position: vscode.Position
+        document: TextDocument,
+        position: Position
     ): Definitions {
         if (!this.uses) throw new Error("Wreq isnt used anywhere");
 
@@ -190,7 +135,7 @@ export class WebpackAstParser {
         const modules = new Map(
             [...this.vars.entries()].filter(([k]) => {
                 return this.uses!.uses.some(e => {
-                    const node = findParrent<VariableDeclaration>(
+                    const node = findParrent(
                         e.location,
                         isVariableDeclaration
                     );
@@ -209,12 +154,11 @@ export class WebpackAstParser {
 
         if (!requiredModule) return;
 
-        const [, dec] =
-            [...modules.entries()].find(([, v]) => {
-                return v.uses.some(({ location }) => requiredModule === location);
-            }) ?? [];
+        const [, dec] = [...modules.entries()].find(([, v]) => {
+            return v.uses.some(({ location }) => requiredModule === location);
+        }) ?? [];
 
-        const moduleId = getModuleId(dec, exportName);
+        const moduleId = getModuleId(dec);
 
         if (!moduleId) return;
         const res = await sendAndGetData<"rawId">({
@@ -233,26 +177,22 @@ export class WebpackAstParser {
         };
     }
 
-    public async generateReferences(document: vscode.TextDocument, position: vscode.Position): References {
+    public async generateReferences(document: TextDocument, position: Position): References {
         if (!this.thisId)
             throw new Error("Could not find module id of module to search for references of");
         const moduleExports = this.getExportMap();
         const where = this.getDeps();
-        const locs: vscode.Location[] = [];
+        const locs: Location[] = [];
 
-        console.log(moduleExports);
         for (const [name] of Object.entries(moduleExports).filter(([, v]) => v.some(x => {
             if (!x) return;
-            console.log(x), console.log(position);
             return x.contains(position);
         }))) {
-            console.log("Fond mod");
             for (const mod of where?.sync ?? []) {
                 const modText = await ModuleCache.getModuleFromNum(mod);
                 if (!modText) continue;
                 const uses = new WebpackAstParser(modText).getUsesOfExport(this.thisId, name);
-                locs.push(...uses.map(x =>
-                    new vscode.Location(ModuleCache.getModuleURI(mod), x)));
+                locs.push(...uses.map(x => new Location(ModuleCache.getModuleURI(mod), x)));
             }
         }
         return locs;
@@ -262,9 +202,9 @@ export class WebpackAstParser {
         return Object.assign({}, this.getExportMapWreq_d() ?? {}, this.getExportMapWreq_t() ?? {}, this.getExportMapWreq_e() ?? {});
     }
 
-    private getUsesOfExport(moduleId: string, exportName: string): vscode.Range[] {
+    private getUsesOfExport(moduleId: string, exportName: string): Range[] {
         if (!this.wreq) throw new Error("Wreq is not used in this file");
-        const uses: vscode.Range[] = [];
+        const uses: Range[] = [];
         for (const { location } of this.vars.get(this.wreq)?.uses ?? []) {
 
             if (!isCallExpression(location.parent)) continue;
@@ -310,7 +250,7 @@ export class WebpackAstParser {
 
         return Object.fromEntries(uses.uses.map(({ location }): [string, ExportMap[string]] | false => {
             const [, exportAssignment] = getLeadingIdentifier(location);
-            const binary = findParrent<BinaryExpression>(location, isBinaryExpression);
+            const binary = findParrent(location, isBinaryExpression);
             if (exportAssignment && binary && isIdentifier(binary?.right)) {
                 return [exportAssignment.text, [makeRange(exportAssignment, this.text), makeRange(binary.right, this.text), this.makeRangeFromFuctionDef(binary.right)]];
             }
@@ -333,7 +273,7 @@ export class WebpackAstParser {
 
         if (!exportAssignment) return undefined;
 
-        const exportObject = findParrent<BinaryExpression | undefined>(
+        const exportObject = findParrent(
             exportAssignment.location, isBinaryExpression
         )?.right;
 
@@ -351,13 +291,13 @@ export class WebpackAstParser {
         if (!wreqD) return;
         const [, exports] = wreqD.arguments;
         return Object.fromEntries(exports.properties.map((x): false | [string, ExportMap[string]] => {
-            if (!isPropertyAssignment(x) || !isFunctionExpression(x.initializer)) return false as const;
+            if (!isPropertyAssignment(x) || !(isArrowFunction(x.initializer) || isFunctionExpression(x.initializer))) return false as const;
             let ret: Node | undefined = findReturnIdentifier(x.initializer);
             ret ??= findReturnPropertyAccessExpression(x.initializer);
             return ret != null ? [x.name.getText(), [makeRange(x.name, this.text), isIdentifier(ret) ? this.makeRangeFromFuctionDef(ret) : undefined]] : false as const;
         }).filter(x => x !== false) as any);
     }
-    private makeRangeFromFuctionDef(ret: Identifier): vscode.Range | undefined {
+    private makeRangeFromFuctionDef(ret: Identifier): Range | undefined {
         const uses = this.vars.get(ret)?.uses;
         if (!uses) return undefined;
         const def = uses.find(({ location }) => isFunctionDeclaration(location.parent));
@@ -365,7 +305,7 @@ export class WebpackAstParser {
         return makeRange(def.location, this.text);
     }
 
-    public findExportLocation(exportName: string): vscode.Range {
+    public findExportLocation(exportName: string): Range {
         return (
             this.tryFindExportwreq_d(exportName) ||
             this.tryFindExportWreq_t(exportName) ||
@@ -375,18 +315,16 @@ export class WebpackAstParser {
     }
     private findWreq_d(): (CallExpression & { arguments: [Identifier, ObjectLiteralExpression, ...any]; }) | undefined {
         if (this.uses) {
-            const maybeWreqD = this.uses.uses.find(use =>
-                getLeadingIdentifier(use.location)[1]?.text === "d"
+            const maybeWreqD = this.uses.uses.find(use => getLeadingIdentifier(use.location)[1]?.text === "d"
             )?.location.parent.parent;
             if (!maybeWreqD || !isCallExpression(maybeWreqD)) return undefined;
             if (maybeWreqD.arguments.length !== 2 ||
                 !isIdentifier(maybeWreqD.arguments[0]) ||
-                !isObjectLiteralExpression(maybeWreqD.arguments[1])
-            ) return undefined;
+                !isObjectLiteralExpression(maybeWreqD.arguments[1])) return undefined;
             return maybeWreqD as any;
         }
     }
-    private tryFindExportwreq_d(exportName: string): vscode.Range | undefined {
+    private tryFindExportwreq_d(exportName: string): Range | undefined {
         if (this.uses) {
             const wreq_dCall = this.findWreq_d();
             if (!wreq_dCall) return undefined;
@@ -397,30 +335,34 @@ export class WebpackAstParser {
                 exportName
             );
 
-            if (
-                !exportCallAssignment ||
+            if (!exportCallAssignment ||
                 !isPropertyAssignment(exportCallAssignment) ||
-                !isFunctionExpression(exportCallAssignment.initializer)
-            )
+                !(isFunctionExpression(exportCallAssignment.initializer)
+                    || isArrowFunction(exportCallAssignment.initializer)))
                 return undefined;
 
             const exportVar = findReturnIdentifier(exportCallAssignment.initializer);
+            if (exportVar) {
 
-            const [exportDec] =
-                [...this.vars.entries()].find(([, v]) => {
+                const [exportDec] = [...this.vars.entries()].find(([, v]) => {
                     return v.uses.some(use => use.location === exportVar);
                 }) ?? [];
 
-            if (!exportDec) return undefined;
+                if (!exportDec) return undefined;
 
-            return makeRange(exportDec, this.text);
+                return makeRange(exportDec, this.text);
+            }
+            const reExport = findReturnPropertyAccessExpression(exportCallAssignment.initializer);
+            if (reExport) {
+                return makeRange(reExport.name, this.text);
+            }
         }
     }
 
     private findWreq_t(): Identifier | undefined {
         return findWebpackArg(this.sourceFile, 1);
     }
-    private tryFindExportWreq_t(exportName: string): vscode.Range | undefined {
+    private tryFindExportWreq_t(exportName: string): Range | undefined {
         const wreq_t = this.findWreq_t();
 
         if (!wreq_t) return undefined;
@@ -439,7 +381,7 @@ export class WebpackAstParser {
     private findWreq_e(): Identifier | undefined {
         return findWebpackArg(this.sourceFile, 0);
     }
-    private tryFindExportsWreq_e(exportName: string): vscode.Range | undefined {
+    private tryFindExportsWreq_e(exportName: string): Range | undefined {
         const wreq_e = this.findWreq_e();
 
         if (!wreq_e) return undefined;
@@ -455,7 +397,7 @@ export class WebpackAstParser {
 
         if (!exportAssignment) return undefined;
 
-        const exportObject = findParrent<BinaryExpression | undefined>(
+        const exportObject = findParrent(
             exportAssignment.location,
             isBinaryExpression
         )?.right;
