@@ -1,6 +1,8 @@
-import { TreeItem, TreeItemCollapsibleState } from "vscode";
+import { TNode } from "@type/sidebar";
 
 import RuntimeCommand from "./RuntimeCommand";
+
+import { TreeItem, TreeItemCollapsibleState } from "vscode";
 
 export class Item extends TreeItem {
     parrent: TNode | null = null;
@@ -41,9 +43,3 @@ export class Section extends Item {
         });
     }
 }
-
-export interface IDynamicNode {
-    getNode(): Promise<Item>;
-}
-export type TNode = Item | IDynamicNode;
-
