@@ -5,10 +5,10 @@
  */
 
 // should be the same types as src/server/types/recieve.ts in the companion
-import { ReporterData as IReporterData } from "../../types";
+import { ReporterData as IReporterData } from "../reporter";
 export type ReporterData = IReporterData;
 
-export type IncomingMessage = (Report | DiffModule | ExtractModule | ModuleList | RawId) & Base;
+export type IncomingMessage = (Report | DiffModule | ExtractModuleR | ModuleList | RawId) & Base;
 export type FullIncomingMessage = IncomingMessage & Nonce;
 export type Base = {
     ok: true;
@@ -37,7 +37,7 @@ export type DiffModule = {
     } & ModuleResult;
 };
 
-export type ExtractModule = {
+export type ExtractModuleR = {
     type: "extract";
     data: {
         module: string;
